use build tags to ignore DSA tests on unsupported golang versions
